requirements.txt
requests==2.31.0
pandas==2.2.0
python-dotenv==1.0.0
entsoe-py==0.6.6
openpyxl==3.1.2
matplotlib==3.8.2
seaborn==0.13.1
streamlit==1.31.0
plotly==5.18.0
